передает в Создание все офферы и чекнутые
